requirements.txt
requests
uvicorn
fastapi
cjdg_api
pandas
pymysql
aiofiles
openpyxl
XlsxWriter
loguru
prettytable
xlwt
xlrd
openpyxl
python-multipart
jinja2
python-jose
scrapy